Add a helper method to fetch all values
This a function rather then an array because we may not necessarily
want to trigger the evaluation of all child attributes.  Having this
be an array would mean that any evaluation of a synthetic attribute
would force evaluation of that attribute for all nodes in the system.
By making it a function, we preserve the lazy evaluation approach.