Adding MobX as a development dependency
For now, I don't want MobX to be an actual dependency.  But I want this to
work with MobX.  So I'm going to see whether I can make the library work
easily with MobX without it being MobX aware or whether I'll have to have
some kind of plugin or extension for that.