Using Object.defineProperty now
This seems like a super clean way to go.  It should work with MobX
without issue.  It is conceptually quite simple.  The addition of
attributes can be done in a built-in way and types can be handled
via some simple casting to add the attributes.  Ideally, the types
will be dealt with by the point where the callbacks are called.
This will allow the evaluation callbacks to reference, in a type
safe way, any _previously defined_ attributes.