I seem to have gotten all the type constraints correct
This was a difficult thing to untangle.  I wanted to express super constraints (ala Java/Scala),
but TypeScript doesn't seem to handle this.  I suspect perhaps a bug in the TypeScript compiler,
but I'm not sure I could provide it.  What I ultimately settled on (which doesn't really impact
the user) is passing the same argument twice, once to satisfy a restricted type and the other
to satisfy a more expansive type.  I'm not sure why this is necessary to get the compiler
to properly check for the things I want it to check for, but it worked and it doesn't make
the code less readable, just a bit puzzling why the extra data is required.