Changed tree definition
To move to attributes as first class properties on nodes, we need
to restrict nodes to being instances of object.  Previously, I
was allowing other types to be nodes.  But with this new approach
where we annotate the tree with new properties, those properties
must be (can only be) added to object instances.  Therefore,
the only things that can have attributes are objects...therefore,
only objects can be nodes in our trees.

This just involved changing a few test cases.