Better understanding of observability requirements
This now handles the simple case of modifying a leaf
in a tree and having it be "observed" via MobX
by computed properties.

The trick now is to handle structural changes.
For that, I suspect I'll need to refactor some of
the tree walking stuff so that it remaps the tree and then
annotates any unannotated notes.