Need a general statistics framework
So I can keep track of things like number of times a given evaluator was
called, number of times per node, number of times an attribute was
invoked, number of times per node, etc.

To do this, I don't want to keep creating wrappers around functions because
they we get an explosing of 'entities' we need to track.  Instead, I think
I'll see if I can inject this into the args used when calling each
evaluator and have a way of wrapping the evaluator to invoke a callback
back to the Arbor instance.  We'll see...