A somewhat half-baked state
There are a couple of things currently going on.  First, I need to
extract better metrics about what is going on (counting evaluations,
testing lazy evaluation, etc).  The goal of the tests is not only to
ensure the proper value is returned, but that it is computed with
a minimal of extra calculations.  Instrumenting this to truly measure
what I want is something I need to sort out still.

The other issue is how plugins are working.  They are related to the issue above.
The functional approach I'm taking, while powerful, tends to obfuscate
what is going on making it difficult for me to measure things.  I need to sort
out how to make the system extensible enough to allow the range of
functionality I want (and extensibility) without being so complex that
I can't instrument the code or reason about it.