Using built-in parent attribute
This optimizes lookup for inherited attributes (substantially).  The idea is that
we have a built-in parent attribute and for all other inherited attributes we
use that to determine parent.  In other words, we are provided with a tree
that is built via-child relations be we reconstruct the parent relations
a the outset so we have both.  This allows for fast traversal both up
and down the tree.