Make parents special
Trying to handle the parent attribute (which is
special because it must be eager and walk
the whole tree first) and other inherited attributes
in the same way was too complicated.  This is
much simpler.

Also, uncovered a nasty but related to interactions
between emitting events and caching results (too late).