Immutable trees DO require cache invalidation...for inherited attributes
I was focusing on synthetic attributes and those are no issue for caching because
an immutable node and its children must have the same synthetic attribute values.
But when that node gets reparented into a new tree, any inherited attributes get
invalidated and I wasn't handling that case until now.